refactor: migrate escrow form state management to useEscrowBoundedStore for improved consistency and remove deprecated store
