+++ src/components/modules/escrow/hooks/initialize-escrow.hook.ts
@@ -3,7 +3,6 @@
 
 import { toast } from "@/hooks/toast.hook";
 import { initializeEscrow } from "@/components/modules/escrow/services/initialize-escrow.service";
-import { useEscrowFormStore } from "@/store/escrowFormStore/store";
 import { zodResolver } from "@hookform/resolvers/zod";
 import { useForm } from "react-hook-form";
 import { useEffect, useMemo } from "react";
@@ -22,7 +21,9 @@ export const useInitializeEscrow = () => {
   const addEscrow = useGlobalBoundedStore((state) => state.addEscrow);
   const loggedUser = useGlobalAuthenticationStore((state) => state.loggedUser);
   const setIsLoading = useGlobalUIBoundedStore((state) => state.setIsLoading);
-  const { formData, setFormData, resetForm } = useEscrowFormStore();
+  const formData = useEscrowBoundedStore((state) => state.formData);
+  const setFormData = useEscrowBoundedStore((state) => state.setFormData);
+  const resetForm = useEscrowBoundedStore((state) => state.resetForm);
   const router = useRouter();
   const setIsSuccessDialogOpen = useEscrowBoundedStore(
     (state) => state.setIsSuccessDialogOpen,
