+++ src/components/modules/escrow/store/ui/index.ts
@@ -8,11 +8,14 @@ import { useEscrowViewModeSlice } from "./slices/view-mode.slice";
 import { ViewModeEscrowStore } from "./@types/view-mode.entity";
 import { useEscrowLoadersSlice } from "./slices/loaders.slice";
 import { LoadersEscrowStore } from "./@types/loaders.entity";
+import { useEscrowInitializeFormSlice } from "./slices/initialize-form.slice";
+import { InitializeFormEscrowStore } from "./@types/initialize-form.entity";
 
 type GlobalState = DialogEscrowStore &
   TabsEscrowStore &
   ViewModeEscrowStore &
-  LoadersEscrowStore;
+  LoadersEscrowStore &
+  InitializeFormEscrowStore;
 
 const devtoolsOptions: DevtoolsOptions = {
   name: "Global State",
@@ -49,6 +52,7 @@ export const useEscrowBoundedStore = create<GlobalState>()(
       ...useEscrowTabSlice(...a),
       ...useEscrowViewModeSlice(...a),
       ...useEscrowLoadersSlice(...a),
+      ...useEscrowInitializeFormSlice(...a),
     }),
     devtoolsOptions,
   ),
