+++ src/components/modules/escrow/store/ui/slices/initialize-form.slice.ts
@@ -1,0 +1,40 @@
+import { StateCreator } from "zustand";
+import {
+  EscrowFormData,
+  InitializeFormEscrowStore,
+} from "../@types/initialize-form.entity";
+
+export const useEscrowInitializeFormSlice: StateCreator<
+  InitializeFormEscrowStore,
+  [["zustand/devtools", never]],
+  [],
+  InitializeFormEscrowStore
+> = (set) => {
+  const initialState: EscrowFormData = {
+    client: "",
+    engagementId: "",
+    serviceProvider: "",
+    platformAddress: "",
+    platformFee: "",
+    amount: "",
+    releaseSigner: "",
+    disputeResolver: "",
+    milestones: [{ description: "" }],
+  };
+
+  return {
+    // Stores
+    ...initialState,
+    formData: initialState,
+
+    // Modifiers
+    setFormData: (data) =>
+      set((state) => ({
+        formData: {
+          ...state.formData,
+          ...data,
+        },
+      })),
+    resetForm: () => set({ formData: initialState }),
+  };
+};
