+++ src/components/ui/steps.tsx
@@ -1,9 +1,10 @@
 "use client";
 
 import * as React from "react";
-import { useStepsStore } from "@/store/stepsStore/store";
 import { cn } from "@/lib/utils";
 import { Check } from "lucide-react";
+import { useEscrowBoundedStore } from "../modules/escrow/store/ui";
+import { useEscrowStepsStore } from "@/store/EscrowStepStore/store";
 
 export interface StepItem {
   title: string;
@@ -18,7 +19,7 @@ export interface StepsProps extends React.HTMLAttributes<HTMLDivElement> {
 export const Steps = React.forwardRef<HTMLDivElement, StepsProps>(
   ({ items, className, ...props }, ref) => {
     const { currentStep, setTotalSteps, toggleStep, isStepCompleted } =
-      useStepsStore();
+      useEscrowStepsStore();
 
     const [progressStates, setProgressStates] = React.useState<number[]>(
       items.map(() => 0),
